\docType{package}
\name{camb}
\alias{camb}
\alias{camb-package}
\title{Chemically Aware Model Builder (camb): An R package for property and bioactivity modeling of small molecules}
\description{
camb allows molecule standardisation, descriptor generation, model building, model ensembling and new molecule
predictions to be done within the same environment.
Two comprehensive tutorials are also provided with the package.
We strongly encourage camb users to read them.
}
\references{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>.
Chemically Aware Model Builder (camb): An R package for property and bioactivity modeling of small molecules.
}

