<<<<<<< HEAD
=======
% Generated by roxygen2 (4.0.0): do not edit by hand
>>>>>>> e07b0708d708c3ba1fc1c5c2123bee363b2657a3
\name{checkModels_extractTypes}
\alias{checkModels_extractTypes}
\title{Check that a list of models are all train objects and are ready to be ensembled together}
\usage{
<<<<<<< HEAD
  checkModels_extractTypes(list_of_models)
}
\arguments{
  \item{list_of_models}{a list of caret models to check}
}
\description{
  Check that a list of models are all train objects and are
  ready to be ensembled together
=======
checkModels_extractTypes(list_of_models)
}
\arguments{
\item{list_of_models}{a list of caret models to check}
}
\description{
Check that a list of models are all train objects and are ready to be ensembled together
>>>>>>> e07b0708d708c3ba1fc1c5c2123bee363b2657a3
}

