<<<<<<< HEAD
=======
% Generated by roxygen2 (4.0.0): do not edit by hand
>>>>>>> e07b0708d708c3ba1fc1c5c2123bee363b2657a3
\name{checkPreds}
\alias{checkPreds}
\title{Check that a list of predictions from caret models are all valid}
\usage{
<<<<<<< HEAD
  checkPreds(list_of_models)
}
\arguments{
  \item{list_of_models}{a list of caret models to check}
}
\description{
  Check that a list of predictions from caret models are
  all valid
=======
checkPreds(list_of_models)
}
\arguments{
\item{list_of_models}{a list of caret models to check}
}
\description{
Check that a list of predictions from caret models are all valid
>>>>>>> e07b0708d708c3ba1fc1c5c2123bee363b2657a3
}

