% Generated by roxygen2 (4.0.1): do not edit by hand
\name{CorrelationPlot}
\alias{CorrelationPlot}
\title{TBD}
\usage{
CorrelationPlot(pred, obs, margin = NULL, main = "", ylab = "Predicted",
  xlab = "Observed", PointSize = 4, ColMargin = "blue", TextSize = 15,
  TitleSize = 15, XAxisSize = 15, YAxisSize = 15, TitleAxesSize = 15,
  tmar = 1, bmar = 1, rmar = 1, lmar = 1, AngleLab = 30,
  LegendPosition = "right", PointColor = "black", PointAlpha = 1,
  PointShape = 16, MarginWidth = 1)
}
\arguments{
\item{tbd}{tbd}
}
\value{
something
}
\description{
TBD description
}
\author{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>
}

