% Generated by roxygen2 (4.0.0): do not edit by hand
\name{DensityResponse}
\alias{DensityResponse}
\title{Visualise the target distribution}
\usage{
DensityResponse(Data, xlab = "", ylab = "", main = "", alpha = 0.2,
  binwidth = NULL, histFill = "white", histCol = "black",
  densityFill = "#FF6666", TitleSize = 15, TextSize = 15,
  XAxisSize = 15, YAxisSize = 15, AngleLab = 30,
  LegendPosition = "right", TitleAxesSize = 15, tmar = 1, bmar = 1,
  rmar = 1, lmar = 1)
}
\arguments{
\item{structures.file}{A character, vector, matrix or data.frame containing the amino acids in either one-letter or three-letter format.
Amino acids symbols are valid in capitals or in lower-case.}
}
\value{
A data.frame with the properties of the molecules read from the original structure file.
}
\description{
TBD: add details here and complete the documentation for the rest of the method (which has another methods docs for now)
}
\examples{
test_mols <- system.file("test_structures", "structures_10.sdf", package = "camb")
StandardiseMolecules(structures.file=test_mols, standardised.file="st.sdf", removed.file="removed.sdf", properties.file="properties.csv")
}
\author{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>
}
\references{
\url{http://www.ggasoftware.com/opensource/indigo}
}

