% Generated by roxygen2 (4.0.1): do not edit by hand
\name{DrawMoleculeInSDF}
\alias{DrawMoleculeInSDF}
\title{2D chemical structure visualization}
\usage{
DrawMoleculeInSDF(structures.file, structure.number, file.name, useNameAsTitle)
}
\description{
DrawMoleculeInSDF permits the depiction of 2D chemical structures.
See the tutorials for examples. 
}
\author{
Isidro Cortes <isidrolauscher@gmail.com> and Daniel Murrell <dsmurrell@gmail.com>
}

