% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ErrorBarplot}
\alias{ErrorBarplot}
\title{TBD}
\usage{
ErrorBarplot(X, Y, err, fill = X, main = "", ylab = "", xlab = "",
  TextSize = 15, TitleSize = 15, XAxisSize = 15, YAxisSize = 15,
  TitleAxesSize = 15, AngleLab = 35, barcol = "red", barSize = 1,
  barWidth = 0.3, LegendName = "Legend", ColLegend = 1,
  RowLegend = NULL, LegendPosition = "right", tmar = 1, bmar = 1,
  rmar = 1, lmar = 1, stat = "identity")
}
\arguments{
\item{tbd}{tbd}
}
\value{
something
}
\description{
TBD description
}
\author{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>
}

