% Generated by roxygen2 (4.0.1): do not edit by hand
\name{expGrid}
\alias{expGrid}
\title{Creates an exponential grid}
\usage{
expGrid(power.from, power.to, power.by, base)
}
\arguments{
\item{power.from}{The first power to use}
}
\description{
Creates an exponential grid
}
\author{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>
}

