<<<<<<< HEAD
=======
% Generated by roxygen2 (4.0.0): do not edit by hand
>>>>>>> e07b0708d708c3ba1fc1c5c2123bee363b2657a3
\name{extractBestPreds}
\alias{extractBestPreds}
\title{Extract predictions for the best tune from a list of caret models}
\usage{
<<<<<<< HEAD
  extractBestPreds(list_of_models)
}
\arguments{
  \item{list_of_models}{a list of caret models to extract
  predictions from}
}
\description{
  Extract predictions for the best tune from a list of
  caret models
=======
extractBestPreds(list_of_models)
}
\arguments{
\item{list_of_models}{a list of caret models to extract predictions from}
}
\description{
Extract predictions for the best tune from a list of caret models
>>>>>>> e07b0708d708c3ba1fc1c5c2123bee363b2657a3
}

