% Generated by roxygen2 (4.0.0): do not edit by hand
\name{GeneratePadelDescriptors}
\alias{GeneratePadelDescriptors}
\title{GeneratePadelDescriptors}
\usage{
GeneratePadelDescriptors(standardised.file, types = c("2D"), threads = -1,
  limit = -1)
}
\description{
TBD: fill in description and the rest
}

