% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetCVTrainControl}
\alias{GetCVTrainControl}
\title{GetCVTrainControl}
\usage{
GetCVTrainControl(ss, seed = 1, folds = 5, repeats = 1, method = "cv",
  returnResamp = "none", returnData = TRUE, savePredictions = TRUE,
  verboseIter = TRUE, allowParallel = TRUE, ...)
}
\arguments{
\item{ss}{the ss}
}
\value{
something
}
\description{
TBD description
}
\author{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>
}

