% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ImputeFeatures}
\alias{ImputeFeatures}
\title{Feature imputation}
\usage{
ImputeFeatures(d, k = 10, ...)
}
\arguments{
\item{d}{A data.frame}

\item{k}{The number of neighbours to use in the imputation}
}
\value{
A data.frame with NA values replaced by imputed values at those locations.
}
\description{
impute.knn is used to impute features using a nearest neighbour based approach.
}
\author{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>
}

