% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MAE}
\alias{MAE}
\title{Mean Absolute Error (MAE)}
\usage{
MAE(v1, v2)
}
\arguments{
\item{v1}{Input vector. e.g. the predicted values for the
	dependent variable.}
\item{v2}{Input vector. e.g. the observed values for the
	dependent variable.}
}
\description{
Mean Absolute Error (MAE) between two numerical input vectors.
}
\author{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>
}

