<<<<<<< HEAD
=======
% Generated by roxygen2 (4.0.0): do not edit by hand
>>>>>>> e07b0708d708c3ba1fc1c5c2123bee363b2657a3
\name{makePredObsMatrix}
\alias{makePredObsMatrix}
\title{Extract obs from one models, and a matrix of predictions from all other models}
\usage{
<<<<<<< HEAD
  makePredObsMatrix(list_of_models)
}
\arguments{
  \item{list_of_models}{a list of caret models to extract
  predictions from}
}
\description{
  Extract obs from one models, and a matrix of predictions
  from all other models
=======
makePredObsMatrix(list_of_models)
}
\arguments{
\item{list_of_models}{a list of caret models to extract predictions from}
}
\description{
Extract obs from one models, and a matrix of predictions from all other models
>>>>>>> e07b0708d708c3ba1fc1c5c2123bee363b2657a3
}

