% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mergeData}
\alias{mergeData}
\title{Merge descriptors blocks.}
\usage{
mergeData(a1, a2, a3, a4 = NULL, a5 = NULL, a6 = NULL)
}
\arguments{
\item{a1..a6}{Descriptors blocks with the same number of rows to be merged.}
}
\value{
The merged block of descriptors.
}
\description{
The function merges blocks of descriptors by columns.
}
\author{
Isidro Cortes <isidrolauscher@gmail.com> and Daniel Murrell <dsmurrell@gmail.com>
}

