% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MinPerf}
\alias{MinPerf}
\title{TBD: fill in description and the rest}
\usage{
MinPerf(meanNoise = 0, sdNoise, resp, lenPred, stds = NULL, iters = 1000,
  filename = NULL, pdfW = 10, pdfH = 10, TextSize = 15,
  TitleSize = 15, XAxisSize = 15, YAxisSize = 15, TitleAxesSize = 15,
  tmar = 1, bmar = 1, rmar = 1, lmar = 1, AngleLab = 30,
  LegendPosition = "right")
}
\description{
TBD: fill in description and the rest
}

