% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PCA}
\alias{PCA}
\title{Principal Component Analysis (PCA)}
\usage{
PCA(Data, RowNames = NULL, cor = TRUE, scale = TRUE, center = TRUE, ...)
}
\description{
The function "PCA" enables the calculation of the Principal Components (PCs) for a given set of descriptors.
The function takes as arguments the descriptors and, optionally, the names of the rows, i.e. datapoints.
Further arguments of the function prcomp from the package stats, use to run the PCA analysis, can be additionally set.
The function returns a list with following elements :

Data : a dataframe containing the two first PCs and the row names if provided. 
Rows are indexed as in the input data corresponds to a list.
PCs_All| : a dataframe containing all PCs.
Std : a vector containing the standard deviation of all PCs.
Info : information about the PCA analysis, such as the proportion of variance explained by each PC.
It is always advisable to verify that the two or three first PCs explain a large proportion of the variance in the data,
if conclusions are to be extracted from this type of analysis.

}

\author{
Isidro Cortes <isidrolauscher@gmail.com> and Daniel Murrell <dsmurrell@gmail.com>
}
