% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PCAPlot}
\alias{PCAPlot}
\title{TBD: fill in description and the rest}
\usage{
PCAPlot(Data, main = "", ylab = "PC2", xlab = "PC1", labels = NULL,
  PointSize = 4, LegendPosition = "right", LegendName = "",
  ColLegend = 1, RowLegend = NULL, TitleSize = 15, TextSize = 15,
  XAxisSize = 15, YAxisSize = 15, AngleLab = 30, TitleAxesSize = 15,
  LegendTitleSize = 15, LegendTextSize = 15, tmar = 1, bmar = 1,
  rmar = 1, lmar = 1)
}
\description{
TBD: fill in description and the rest
}

