<<<<<<< HEAD
=======
% Generated by roxygen2 (4.0.0): do not edit by hand
>>>>>>> e07b0708d708c3ba1fc1c5c2123bee363b2657a3
\name{predict.caretStack}
\alias{predict.caretStack}
\title{Make predictions from a caretStack. This function passes the data to each function in
turn to make a matrix of predictions, and then multiplies that matrix by the vector of
weights to get a single, combined vector of predictions.}
\usage{
<<<<<<< HEAD
  predict.caretStack(ensemble, newdata = NULL, ...)
}
\arguments{
  \item{ensemble}{a caretStack to make predictions from.}

  \item{...}{arguments (including newdata) to pass to
  predict.train.}
}
\description{
  Make predictions from a caretStack. This function passes
  the data to each function in turn to make a matrix of
  predictions, and then multiplies that matrix by the
  vector of weights to get a single, combined vector of
  predictions.
=======
predict.caretStack(ensemble, newdata = NULL, ...)
}
\arguments{
\item{ensemble}{a caretStack to make predictions from.}

\item{...}{arguments (including newdata) to pass to predict.train.}
}
\description{
Make predictions from a caretStack. This function passes the data to each function in
turn to make a matrix of predictions, and then multiplies that matrix by the vector of
weights to get a single, combined vector of predictions.
>>>>>>> e07b0708d708c3ba1fc1c5c2123bee363b2657a3
}

