% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PredictExternal}
\alias{PredictExternal}
\title{Make predictions on new molcules using a saved 
	standardisation procedure and a saved model}
\usage{
PredictExternal(structures.file, standardisation.options, descriptor.types,
  dataset, model)
}
\arguments{
\item{structures.file}{The name of the file containing the chemical structures. 
	SMILES and SDF are currently supported formats.}

\item{standardisation.options}{The options saved during the standardisation procedure. 
	These options are returned from the StandardiseMolecules function.}

\item{descriptor.types}{A named list of the types of descriptors used in model training.}

\item{dataset}{The dataset used in model training. This is used for the preprocessing 
	applied to the training data as well as the descriptors used in training.}

\item{model}{The trained model.}
}
\value{
A data.frame containing the original ids of the molecules as well as their predicted values.
}
\description{
Molecules are converted to a standard representation in the same way as during model training.
A saved model is used to make predictions on new molecules.
}
\examples{
#test_structures_file <- system.file("test_structures", "structures_10.sdf", package = "camb")
# The following requires a trained model, e.g. 'rf.rds'
#predictions <- PredictExternal(test_structures_file, 
#standardisation.options, descriptor.types, dataset, readRDS("rf.rds"))
}
\author{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>
}

