% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PreProcess}
\alias{PreProcess}
\title{TBD}
\usage{
PreProcess(ss, steps = c("center", "scale"), ...)
}
\arguments{
\item{ss}{the ss}
}
\value{
something
}
\description{
TBD description
}
\author{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>
}

