% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Qsquared1}
\alias{Qsquared1}
\title{TBD}
\usage{
Qsquared1(v1, v2, resp_tr)
}
\arguments{
\item{tbd}{tbd}
}
\value{
something
}
\description{
TBD description
}
\author{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>
}

