% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Qsquared3}
\alias{Qsquared3}
\title{TBD}
\usage{
Qsquared3(v1, v2, resp_tr)
}
\arguments{
\item{tbd}{tbd}
}
\value{
something
}
\description{
TBD description
}
\author{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>
}

