% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RemoveHighlyCorrelatedFeatures}
\alias{RemoveHighlyCorrelatedFeatures}
\title{Removal of highly correlated features}
\usage{
RemoveHighlyCorrelatedFeatures(dataset, correlationCutoff = 0.95, ...)
}
\arguments{
\item{dataset}{An already split dataset (training/holdout) returned by the SplitSet function.}

\item{correlationCutoff}{The correlation threshold.}
}
\value{
A split dataset list with any pairs of correlated descriptors with correlation above the threshold
\code{correlationCutoff} removed from both the training and holdout set.
}
\description{
Highly correlated features are removed. If two features have a Pearson's correlation above \code{correlationCutoff}
then the one with the highest mean correlation to the remaining features.
}
\author{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>
}

