% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RemoveNearZeroVarianceFeatures}
\alias{RemoveNearZeroVarianceFeatures}
\title{Removal of near zero variance descriptors}
\usage{
RemoveNearZeroVarianceFeatures(dataset, frequencyCutoff = 30/1, ...)
}
\arguments{
\item{dataset}{An already split dataset (training/holdout) returned by the SplitSet function.}

\item{frequencyCutoff}{The frequency ratio threshold.}
}
\value{
A split dataset list with near zero variance descriptors removed from both the training and holdout set.
}
\description{
Near variance descriptors are removed by checking the ratio of the frequency of the most commonly occuring value
to the frequency of the second most commonly occuring value. If this ratio exceeds the threshold \code{frequencyCutoff}, then
the feature is removed.
}
\author{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>
}

