% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RemoveStandardisedPrefix}
\alias{RemoveStandardisedPrefix}
\title{Remove the prefix that was added to the molecule names before standardisation.}
\usage{
RemoveStandardisedPrefix(descriptors)
}
\arguments{
\item{descriptors}{A \code{data.frame} containing the descriptors.}
}
\value{
A \code{data.frame} containing the descriptors with the prefix removed.
}
\description{
This function is needed to remove the prefix that is added to the molecule names to make sure that they don't start with a number.
PaDEL-Descriptor doesn't handle molecule names starting with certain characters so a prefix is added before its use and then removed with this function.
This function will be removed at some point when the PaDEL-Descriptor issue is resolved.
}

