% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ReplaceInfinitesWithNA}
\alias{ReplaceInfinitesWithNA}
\title{Remove infinite values from the descriptor data.frame}
\usage{
ReplaceInfinitesWithNA(d)
}
\arguments{
\item{d}{A data.frame}
}
\value{
A data.frame with infinite values replaced by NA.
}
\description{
Any infinites found in the descriptor data.frame are replaced with NA.
}
\author{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>
}

