\name{RMSE}
\alias{RMSE}
\title{RMSE}
\usage{
RMSE(v1, v2)
}
\arguments{
\item{v1}{Input vector. e.g. the predicted values for the
	dependent variable.}
\item{v2}{Input vector. e.g. the observed values for the
	dependent variable.}
}

\description{
Calculates the RMSE for two input vectors, 
e.g. the observed and the predicted values for a test set.
See the function Validation and the camb tutorials for further information.
}
\author{
Isidro Cortes <isidrolauscher@gmail.com> and Daniel Murrell <dsmurrell@gmail.com>
}

