% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Rsquared0}
\alias{Rsquared0}
\title{TBD}
\usage{
Rsquared0(v1, v2)
}
\arguments{
\item{tbd}{tbd}
}
\value{
something
}
\description{
TBD description
}
\author{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>
}

