% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Rsquared_CV}
\alias{Rsquared_CV}
\title{Calculate the cross validated RSquared}
\usage{
Rsquared_CV(model, digits = 3)
}
\description{
Calculate the cross validated RSquared
}
\author{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>
}

