% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SeqDescs}
\alias{SeqDescs}
\title{Whole Protein Sequence Descriptor Calculation}
\usage{
SeqDescs(data, UniProtID = TRUE, type = "AAC", ...)
}
\arguments{
\item{data}{One or more protein sequences, or one or several UniProt IDs.}

\item{UniProtID}{If TRUE the argument calculates the descriptors for the proteins
which UniProt IDs have been indicated in the argument 'data'.}

\item{type}{The type of protein descriptors to be calculated (see above).
Any combination thereof is valid. A vector containing the abbreviation of the desired descriptors
is taken as argument. Default value 'AAC'.}
}
\value{
A numeric matrix which rows are indexed by proteins and the columns by descriptors.
If multiple descriptors are chosen, the function returns a matrix where descriptors are concatenated per row
for the ease of modeling.
}
\description{
Calculation of the following 12 whole sequence protein descriptors:
Amino Acid Composition ("AAC"), 
Dipeptide Composition ("DC"), 
Tripeptide Composition ("TC"), 
Normalized Moreau-Broto Autocorrelation ("MoreauBroto"), 
Moran Autocorrelation ("Moran"), 
Geary Autocorrelation ("Geary") ,
CTD (Composition/Transition/Distribution) ("CTD"), 
Conjoint Traid ("CTriad"), 
Sequence Order Coupling Number ("SOCN"), 
Quasi-sequence Order Descriptors ("QSO"), 
Pseudo Amino Acid Composition ("PACC"), 
Amphiphilic Pseudo Amino Acid Composition ("APAAC").

}
\author{
Isidro Cortes <isidrolauscher@gmail.com> and Daniel Murrell <dsmurrell@gmail.com>
}
\references{
R package protr}

