% Generated by roxygen2 (4.0.1): do not edit by hand
\name{slope}
\alias{slope}
\title{Slope}
\usage{
slope(v1, v2)
}
\arguments{
\item{v1}{Input vector. e.g. the predicted values for the
	dependent variable.}
\item{v2}{Input vector. e.g. the observed values for the
	dependent variable.}

}
\description{
Slope between two vectors calculated as:
sum(v2 * v1)/sum(v1 * v1)
See the camb tutorials for further information.}
\author{
Isidro Cortes <isidrolauscher@gmail.com> and Daniel Murrell <dsmurrell@gmail.com>
}

