% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SplitSet}
\alias{SplitSet}
\title{Splitting the data into training and holdout set}
\usage{
SplitSet(ids, x, y, percentage = 20, seed = 1)
}
\arguments{
\item{ids}{The ids of the molecules.}

\item{x}{The descriptors data.frame.}

\item{y}{The target values.}

\item{percentage}{The percentage of molecules that are assigned to the holdout set.}

\item{seed}{The seed for the random number generator used to split the data. Using the same seed will result in the same split.}
}
\value{
A list cointaining the training indexes (\code{train.indexes}),
the training descriptors data.frame (\code{x.train}), the training target values (\code{y.train} and
the holdout indexes (\code{holdout.indexes}), the holdout descriptors data.frame (\code{x.holdout}),
and the holdout target values (\code{y.holdout}).
}
\description{
The data is randomly split into a training set and a holdout set of \code{x} percent of the original data.
The data along with the partition are returned as a list which is passed in and out of functions latter down the camb workflow.
}
\author{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>
}

