% Generated by roxygen2 (4.0.0): do not edit by hand
\name{StandardiseMolecules}
\alias{StandardiseMolecules}
\title{Convert molecules to a standard representation}
\usage{
StandardiseMolecules(structures.file, standardised.file, removed.file = "",
  properties.file = "standardisation_info.csv", remove.inorganic = FALSE,
  fluorine.limit = -1, chlorine.limit = -1, bromine.limit = -1,
  iodine.limit = -1, min.mass.limit = -1, max.mass.limit = -1,
  number.processed = -1)
}
\arguments{
\item{structures.file}{A character, vector, matrix or data.frame containing the amino acids in either one-letter or three-letter format.
Amino acids symbols are valid in capitals or in lower-case.}
}
\value{
A data.frame with the properties of the molecules read from the original structure file.
}
\description{
Details about the standardisation procedure
}
\examples{
test_mols <- system.file("test_structures", "structures_10.sdf", package = "camb")
StandardiseMolecules(structures.file=test_mols, standardised.file="st.sdf", removed.file="removed.sdf", properties.file="properties.csv")
}
\author{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>
}
\references{
\url{http://www.ggasoftware.com/opensource/indigo}
}

