% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Validation}
\alias{Validation}
\title{TBD}
\usage{
Validation(pred, obs, resp_tr)
}
\arguments{
\item{tbd}{tbd}
}
\value{
something
}
\description{
TBD description
}
\author{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>
}

