% Generated by roxygen2 (4.0.1): do not edit by hand
\name{YScrambling}
\alias{YScrambling}
\title{TBD}
\usage{
YScrambling(y, percent)
}
\arguments{
\item{power.from}{the from power}
}
\value{
something
}
\description{
TBD description
}
\author{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>
}

