% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{camb}
\alias{camb}
\alias{camb-package}
\title{Chemically Aware Model Builder (camb): An R package for property and bioactivity modeling of small molecules}
\description{
camb allows molecule standardisation, descriptor generation, model building, model ensembling and new molecule
predictions to be done within the same environment.
}
\references{
Daniel Murrell, Isidro Cortes (2014). Chemically Aware Model Builder (camb): An R package for property and bioactivity modeling of small molecules.
}

