% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkModels_extractTypes}
\alias{checkModels_extractTypes}
\title{Check that a list of models are all train objects and are ready to be ensembled together}
\usage{
checkModels_extractTypes(list_of_models)
}
\description{
Check that a list of models are all train objects and are ready to be ensembled together
}
\author{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>
}

