% Generated by roxygen2 (4.0.1): do not edit by hand
\name{expGrid}
\alias{expGrid}
\title{TBD}
\usage{
expGrid(power.from, power.to, power.by, base)
}
\arguments{
\item{power.from}{the from power}
}
\value{
something
}
\description{
TBD description
}
\author{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>
}

