% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GeneratePadelDescriptors}
\alias{GeneratePadelDescriptors}
\title{GeneratePadelDescriptors}
\usage{
GeneratePadelDescriptors(standardised.file, types = c("2D"), threads = -1,
  limit = -1)
}
\arguments{
\item{standardised.file}{The name of the file to which the standardised molecules were written to with the \code{\link{StandardiseMolecules}} function.
If standardisation is not used then this can be any file in the SDF format.}

\item{types}{The types of descriptors calculated. Options include: "2D", "Fingerprinter", "ExtendedFingerprinter", "EStateFingerprinter", "GraphOnlyFingerprinter",
	"MACCSFingerprinter,"PubchemFingerprinter","SubstructureFingerprinter",
	"SubstructureFingerprintCount","KlekotaRothFingerprinter" and "KlekotaRothFingerprintCount"}
}
\value{
A \code{data.frame} containing the descriptors with the prefix removed.
}
\description{
Utilises the PaDEL-Descriptor Java library to generate molecular descriptors.
}

