<<<<<<< HEAD
=======
% Generated by roxygen2 (4.0.0): do not edit by hand
>>>>>>> e07b0708d708c3ba1fc1c5c2123bee363b2657a3
\name{greedOptAUC}
\alias{greedOptAUC}
\title{TODO}
\usage{
<<<<<<< HEAD
  greedOptAUC(X, Y, iter = 100L)
}
\arguments{
  \item{X}{}

  \item{Y}{}

  \item{iter}{}
}
\description{
  TODO
=======
greedOptAUC(X, Y, iter = 100L)
}
\arguments{
\item{X}{}

\item{Y}{}

\item{iter}{}
}
\description{
TODO
>>>>>>> e07b0708d708c3ba1fc1c5c2123bee363b2657a3
}

