% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ImputeFeatures}
\alias{ImputeFeatures}
\title{Impute missing descriptor values using knn.impute}
\usage{
ImputeFeatures(d, k = 10, ...)
}
\arguments{
\item{d}{A data.frame}
}
\value{
A data.frame with infinite values replaced by NA.
}
\description{
A nearest neighbour based
}
\author{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>
}

