% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MorganFPs}
\alias{MorganFPs}
\title{Circular Morgan Fingerprints as specified in RDkit}
\usage{
MorganFPs(bits = 512, radii = c(0,1,2), mols, output, keep = "hashed_binary", 
    images = FALSE, unhashed = FALSE, RDkitPath = "/usr/local/share/RDKit", 
    PythonPath = "/usr/local/lib/python2.7/site-packages", extMols = FALSE, 
   unhashedExt = FALSE, logFile = FALSE) 

}
\arguments{
\item{bits}{Number of bits of the hashed fingerprints.}

\item{radii}{Radii to be considered for the calculation of the fingerprints. Note that a radius of 2 is equivalent to ECFP-4 fingerprints.}

\item{mols}{File containing the input molecules.}

\item{output}{Labels that will be appended to all ouput files.}

\item{keep}{The fingeprints that will be kept after the calculation.
Apart from calculating different types of fingerprints,
the function returns a data.frame with the type of fingerprints specified here.
Possible types are: hashed_binary, hashed_counts, unhashed_binary, unhashed_counts,
hashed_binaryEXT, hashed_countsEXT, unhashed_binaryEXT and unhashed_countsEXT.}

\item{images}{If TRUE individual .pdf files containing the image of each substructure present in the input file,
and for each molecule, are created. Be aware that the number of fingerprints can be large depending on the number
and diversity of the molecules present in the input file. Thus, allow for sufficient memory.}

\item{unhashed}{If TRUE, unhashed fingeprints -both in binary format and with counts- are calculated.}

\item{RDkitPath}{The path to the folder containing the RDkit library in your computer.
On mac, the is equal to the environment variable $RDBASE.}

\item{PythonPath}{Path to python ($PYTHONPATH).}

\item{extMols}{File containing the molecules for which unhashed fingeprints are to be calculated
with respect to the pool of substructures in the molecules present in the file specified in 'mols'.}

\item{unhashedExt}{If TRUE, unhashed fingerprints are calcualted for the molecules in 'extMols'.}

\item{logFile}{File where the log messages will be dropped.}
}
\value{
In the working directory, .csv files will be created containing the different fingeprint types
specified with the function arguments. By default, hashed fingerprint in binary format and with counts
will be created. In addition, the function returns a data.frame with the fingerprint types defined in
the argument 'keep'. In the data.frame, rows are indexed by the molecules in the file containing the
molecules, and columns by the bits in the fingerprint vector.
}
\description{
The function calculates circular Morgan fingerprints for chemical compounds
using the RDkit python library (Greg Landrum).
Hashed fingeprints are calculated in binary format or with counts.
In addition, it also calcualates unhashed fingerprints, both binary and with counts.

}
\examples{
test_mols <- system.file("test_structures", "structures_10.sdf", package = "camb")
MorganFPs(bits=28,radii=c(0,1,2),mols=test_mols,output="test_mols")

#See the camb tutorials for more examples.

}
\author{
Isidro Cortes <isidrolauscher@gmail.com> and Daniel Murrell <dsmurrell@gmail.com>
}
\references{
FingeprintCalculator.py. Isidro Cortes. 2013/2014. \url{http://github.com/isidroc/FingerprintCalculator}.
}

