% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PCAPlot}
\alias{PCAPlot}
\title{Plot PCA analysis}
\usage{
PCAPlot(Data, labels = NULL, Colors = NULL, Shapes = NULL,
  main = "", ylab = "PC2", xlab = "PC1", 
  PointSize = 4, LegendPosition = "right", LegendName = "",
  ColLegend = 1, RowLegend = NULL, TitleSize = 15, TextSize = 15,
  XAxisSize = 15, YAxisSize = 15, AngleLab = 0, TitleAxesSize = 15,
  LegendTitleSize = 15, LegendTextSize = 15, tmar = 1, bmar = 1,
  rmar = 1, lmar = 1)
}
\description{
The function "PCAPlot" provides an easy way to plot the first two PC. 
As all the plotting function provided with camb, it is based on 
ggplot2, which allows further customization by the user.
Below is an example of how to use these two function to do a PCA analysis of the target space, 
which in this case is quantified by the amino acid descriptors of the amino acids present in the binding site
of mammal cyclooxygenases.
}

