% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PreProcess}
\alias{PreProcess}
\title{Cals the \code{\link{preProcess}} function of the caret package which handles data transformation before training.}
\usage{
PreProcess(dataset, steps = c('center', 'scale'), ...)
}
\arguments{
\item{dataset}{The training dataset returned by \code{\link{SplitSet}}}

\item{steps}{A character vector specifying the type of processing. Possible values are 'BoxCox', 'YeoJohnson',
'expoTrans', 'center', 'scale', 'range', 'knnImpute', 'bagImpute', 'medianImpute', pca', 'ica' and 'spatialSign'
for details see the \code{\link{preProcess}} function for more details.}
}
\value{
A dataset with the preprocessed training and holdout set that also stores the transformation
}
\description{
Pre-processing transformation (centering, scaling etc.) can be estimated from the training data and applied to
any data set with the same variables. The simplest form of this is to center and scale all the variables so that
each of their means are 0 and each of their standard devitations is 1. See \code{\link{preProcess}} arguments for
more control.
}
\author{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>
}

