\name{Qsquared3}
\alias{Qsquared3}
\title{Calculates the Q Squared 3}
\usage{
Qsquared3(v1, v2, resp_tr)
}
\arguments{
\item{v1}{Input vector. e.g. the predicted values for the
	dependent variable.}
\item{v2}{Input vector. e.g. the observed values for the
	dependent variable.}

\item{resp_tr}{Vector containing the values of the dependent variable
	corresponding to the datapoints in the training set.}
}
\description{
Calculates the Q squared 3 for two input vectors, 
e.g. the observed and the predicted values for a test set.
See the function Validation and the camb tutorials for further information.}
\author{
Isidro Cortes <isidrolauscher@gmail.com> and Daniel Murrell <dsmurrell@gmail.com>
}

