% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RemoveHighlyCorrelatedFeatures}
\alias{RemoveHighlyCorrelatedFeatures}
\title{Calls the \code{\link{findCorrelation}} function of the caret package which finds the highly correlated descriptors
and removes them from the training and holdout sets.}
\usage{
RemoveHighlyCorrelatedFeatures(dataset, correlationCutoff = 0.95, ...)
}
\arguments{
\item{dataset}{The training dataset returned by \code{\link{SplitSet}}}

\item{correlationCutoff}{A numeric value for the pair-wise absolute correlation cutoff}
}
\value{
A dataset with the appropriate columns cut out of the x.train and x.holdout dataframes
}
\description{
The absolute values of pair-wise correlations are considered. If two variables have a high correlation,
the function looks at the mean absolute correlation of each variable and removes the variable with the
largest mean absolute correlation.
}
\author{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>
}

