% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RemoveHighlyCorrelatedFeatures}
\alias{RemoveHighlyCorrelatedFeatures}
\title{TBD}
\usage{
RemoveHighlyCorrelatedFeatures(ss, correlationCutoff = 0.95, ...)
}
\arguments{
\item{ss}{the ss}
}
\value{
something
}
\description{
TBD description
}
\author{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>
}

