% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RemoveNearZeroVarianceFeatures}
\alias{RemoveNearZeroVarianceFeatures}
\title{Calls the \code{\link{nearZeroVar}} function of the caret package and removes descriptos with near zero variance in the
appropriate columns from the training and holdout sets.}
\usage{
RemoveNearZeroVarianceFeatures(dataset, frequencyCutoff = 30/1, ...)
}
\arguments{
\item{dataset}{The training dataset returned by \code{\link{SplitSet}}}

\item{frequencyCutoff}{The cutoff for the ratio of the most common value to the second most common value}
}
\value{
A dataset with the appropriate columns cut out of the x.train and x.holdout dataframes
}
\description{
\code{\link{nearZeroVar}} diagnoses predictors that have one unique value (i.e. are zero variance predictors) or predictors
that are have both of the following characteristics: they have very few unique values relative to the number
of samples and the ratio of the frequency of the most common value to the frequency of the second most common
value is large. checkConditionalX looks at the distribution of the columns of x conditioned on the levels of
y and identifies columns of x that are sparse within groups of y.
}
\author{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>
}

