% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RMSE_CV}
\alias{RMSE_CV}
\title{Extract the cross validated RMSE from a caret model}
\usage{
RMSE_CV(model, digits = 3)
}
\description{
Extract the cross validated RMSE from a caret model
}
\author{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>
}

