% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Rsquared0}
\alias{Rsquared0}
\title{Rsquared0}
\usage{
Rsquared0(v1, v2)
}
\arguments{
\item{v1}{Input vector. e.g. the predicted values for the
	dependent variable.}
\item{v2}{Input vector. e.g. the observed values for the
	dependent variable.}
}

\description{
Calculates the R squared 0 for two input vectors, 
e.g. the observed and the predicted values for a test set.
See the function Validation and the camb tutorials for further information.
}
\author{
Isidro Cortes <isidrolauscher@gmail.com> and Daniel Murrell <dsmurrell@gmail.com>
}


