% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Rsquared_CV}
\alias{Rsquared_CV}
\title{Calculates the cross validated RSquared from a caret model
}
\usage{
Rsquared_CV(model, digits = 3)
}
\arguments{
\item{model}{
	Trained caret model from which the cross-validated R squared is to be obtained.
	}
	\item{digits}{
		Number of decimals to which the value is to be rounded.
		}
		}


\description{
Calculate the cross validated RSquared
}
\author{
Isidro Cortes <isidrolauscher@gmail.com> and Daniel Murrell <dsmurrell@gmail.com>
}

