% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Rsquared_CV}
\alias{Rsquared_CV}
\title{TBD}
\usage{
Rsquared_CV(model, digits = 3)
}
\arguments{
\item{tbd}{tbd}
}
\value{
something
}
\description{
TBD description
}
\author{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>
}

