% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SplitSet}
\alias{SplitSet}
\title{Split into training and hold out data}
\usage{
SplitSet(ids, x, y, percentage = 20, seed = 1)
}
\arguments{
\item{ids}{The names of the molecules}

\item{x}{The descriptors}

\item{y}{The target values}

\item{percentage}{The percentage of data to put into the holdout set}

\item{seed}{The seed for randomization so that the split can be does in a repeatable way if required}
}
\value{
a list designed to be passed between functions in the camb workflow containing the variables required for training:
ids, holdout.indexes, train.indexes, x.train, x.holdout, y.train, y.holdout
}
\description{
Creates a training/holdout split for the data
}
\author{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>
}

