% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SplitSet}
\alias{SplitSet}
\title{Split into training and hold out data TBD}
\usage{
SplitSet(ids, x, y, percentage = 20, seed = 1)
}
\arguments{
\item{ids}{the ids}
}
\value{
something
}
\description{
description
}
\author{
Daniel Murrell <dsmurrell@gmail.com> and Isidro Cortes <isidrolauscher@gmail.com>
}

