% Generated by roxygen2 (4.0.1): do not edit by hand
\name{StandardiseMolecules}
\alias{StandardiseMolecules}
\title{Convert molecules to a standard representation}
\usage{
StandardiseMolecules(structures.file, standardised.file, removed.file = "",
  properties.file = "standardisation_info.csv", remove.inorganic = FALSE,
  fluorine.limit = -1, chlorine.limit = -1, bromine.limit = -1,
  iodine.limit = -1, min.mass.limit = -1, max.mass.limit = -1,
  number.processed = -1)
}
\arguments{
\item{structures.file}{The name of the file containing the chemical structures. SMILES and SDF are currently supported formats.}

\item{standardised.file}{The name of the file to which the standardised molecules are written to. This file is saved in the SDF format.}

\item{removed.file}{The name of the file to which the standardised molecules that were removed by the filters are written to. This file is saved in SDF format.
If left out, this file is not created.}

\item{properties.file}{The name of the file to which the standardization information is written to. The first column contains the molecules IDs, whereas the second, i.e. Kept, indicated whether the molecule was kept (==1) or not (==0).}

\item{remove.inorganic}{If set \code{TRUE}, molecules that contain any atoms not in {H, C, N, O, P, S, F, Cl, Br, I} are excluded.}

\item{fluorine.limit}{If specified, molecules with more than \code{flourine.limit} Flourine atoms are excluded.}

\item{chlorine.limit}{If specified, molecules with more than \code{chlorine.limit} Chlorine atoms are excluded.}

\item{bromine.limit}{If specified, molecules with more than \code{bromine.limit} Bromine atoms are excluded.}

\item{iodine.limit}{If specified, molecules with more than \code{iodine.limit} Iodine atoms are excluded.}

\item{min.mass.limit}{If specified, molecules with a molecular mass smaller than \code{min.mass.limit} are excluded.}

\item{max.mass.limit}{If specified, molecules with a molecule mass greater than \code{max.mass.limit} are excluded.}

\item{number.processed}{If specified, only the first \code{number.processed} molecules will be processed by this function.
This is used mainly for testing purposed on files that contain a lot of molecules.}
}
\value{
The options used in standardisation so that they may be applied to make predictions on new molecules using the function \code{\link{PredictExternal}}.
}
\description{
Molecules are converted to a standard representation using Indigo's C API. Molecules can be read in either SMILES or SDF format. Hydrogens are made implicit.
Molecules are excluded if they don't pass Indigo's checks for correctness which include incorrect valence representations and ambiguous Hydrogen representations.
Atomic isotopes are converted to their common forms. Molecules are dearomatized and then converted to InChI format using Indigo's InChI plugin.
Molecules are then converted back to a SMILES representation.
Passing them through the InChI format essentially convert all tautomeric forms of the same molecule to a single representation.
Various parameters are available to control which molecule get kept in the standardised set.
}
\examples{
test_mols <- system.file("test_structures", 
						 "structures_10.sdf", package = "camb")
StandardiseMolecules(structures.file=test_mols, 
					 standardised.file="st.sdf", removed.file="removed.sdf", 
					 properties.file="standardization_info.csv")
}
\author{
Isidro Cortes <isidrolauscher@gmail.com> and Daniel Murrell <dsmurrell@gmail.com>
}

\references{
\url{http://www.ggasoftware.com/opensource/indigo}

\url{http://www.iupac.org/home/publications/e-resources/inchi.html}
}

