% Generated by roxygen2 (4.0.1): do not edit by hand
\name{YScrambling}
\alias{YScrambling}
\title{Y-Scrambling}
\usage{
YScrambling(y, percent)
}
\arguments{
\item{y}{Input numerical vector.}
\item{percent}{Percentage of the input vector
	to be randomized.}
}
\value{
A numerical vector with a user-defined percentage of values randomized.
}
\description{
This function serves to randomize (permutate) a subset of the dependent variable.
It is useful to perform Y-scrambling experiments. }
\author{
Isidro Cortes <isidrolauscher@gmail.com> and Daniel Murrell <dsmurrell@gmail.com>
}


